requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.16.2
numpy==1.26.4
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pydantic==2.5.0
python-json-logger==2.0.7
